requirements.txt
absl-py==0.13.0
alabaster==0.7.12
astroid==2.6.5
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
ConfigArgParse==1.5.1
cycler==0.10.0
Cython==0.29.24
dicttoxml==1.7.4
docutils==0.17.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.1
google-auth-oauthlib==0.4.4
grpcio==1.39.0
humanize==3.10.0
idna==3.2
imageio==2.9.0
imagesize==1.2.0
-e git+git@code.cs.uky.edu:seales-research/ink-id.git@36e8542d38baa53326e5ecfe712ad0d7a133bc10#egg=inkid
isort==5.9.2
Jinja2==3.0.1
joblib==1.0.1
jsmin==2.2.2
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mathutils==2.81.2
matplotlib==3.4.2
mccabe==0.6.1
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pygifsicle==1.0.4
Pygments==2.9.0
pylint==2.9.5
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.0
torch-summary==1.4.5
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Wand==0.6.6
Werkzeug==2.0.1
wrapt==1.12.1
